pyproject.toml
[project]
name = "interact-mcp"
version = "0.1.0"
description = "MCP server for browser interaction with vision analysis"
readme = "README.md"
authors = [
    { name = "Alan Blanchet", email = "alan@alan-blanchet.fr" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "litellm>=1.83.4",
    "mcp[cli]>=1.27.0",
    "pillow>=12.2.0",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
artifacts = ["src/interact_mcp/js/*.js"]

[project.scripts]
interact-mcp = "interact_mcp.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests that require external API calls and a real browser",
]
